completes refactoring after basic authentication development completed
