+++ api/utils/security/password_utils.py
@@ -1,5 +1,6 @@
 import bcrypt
 
+
 def hash_password(password):
     pwd_bytes = password.encode('utf-8')
     salt = bcrypt.gensalt()
@@ -10,3 +11,4 @@ def hash_password(password):
 def verify_password(plain_password, hashed_password):
     password_byte_enc = plain_password.encode('utf-8')
     return bcrypt.checkpw(password=password_byte_enc, hashed_password=hashed_password)
+
